Simulate the billiard ball until initial contact
I have added a definition that is supposed to simulate the ball moving
on the table and interacting with the wall. There are errors that only
allow the definition to run until it makes the first contact with the
wall then it stops. It then graphs the balls path along with the table.